src/app/weather/services/weather.service.spec.ts

import { WeatherService } from './weather.service';
import { defer } from 'rxjs';
import { mockedWeatherData } from 'src/app/testing/mock';

function asyncData<T>(data: T) {
  return defer(() => Promise.resolve(data));
}

describe('WeatherService', () => {
  let service: WeatherService;
  let httpClientSpy: { get: jasmine.Spy };

  beforeEach(() => {
    httpClientSpy = jasmine.createSpyObj('HttpClient', ['get']);
    service = new WeatherService(<any> httpClientSpy);
  });

  it('should be created', () => {
    expect(service).toBeTruthy();
  });

  it('should return weather data on success and httpClient should be called once', () => {
    httpClientSpy.get.and.returnValue(asyncData(mockedWeatherData));

    service.searchWeatherForCity('Rome')
      .subscribe(cityWeather => {
        expect(cityWeather).toEqual(mockedWeatherData);
      });

    expect(httpClientSpy.get.calls.count()).toBe(1);
  });
});
